+++ observability/api/api/third.d.ts
@@ -1,0 +1,6 @@
+declare namespace Express {
+  interface Request {
+    user?: import("@blocklet/sdk/lib/util/login").SessionUser;
+    isBlocklet?: boolean;
+  }
+}