+++ packages/cli/src/utils/run-with-aigne.ts
@@ -44,6 +44,7 @@ export interface RunAIGNECommandOptions {
   input?: string[];
   format?: "text" | "json" | "yaml";
   output?: string;
+  outputKey?: string;
   logLevel?: LogLevel;
   force?: boolean;
 }
@@ -56,7 +57,9 @@ export const createRunAIGNECommand = (name = "run") =>
     .option("--chat", "Run chat loop in terminal", false)
     .option(
       "--model <provider[:model]>",
-      `AI model to use in format 'provider[:model]' where model is optional. Examples: 'openai' or 'openai:gpt-4o-mini'. Available providers: ${availableModels.map((i) => i.name.toLowerCase().replace(/ChatModel$/i, "")).join(", ")} (default: openai)`,
+      `AI model to use in format 'provider[:model]' where model is optional. Examples: 'openai' or 'openai:gpt-4o-mini'. Available providers: ${availableModels()
+        .map((i) => i.name.toLowerCase().replace(/ChatModel$/i, ""))
+        .join(", ")} (default: openai)`,
     )
     .option(
       "--temperature <temperature>",
@@ -179,7 +182,8 @@ export async function parseAgentInputByCommander(
 }
 
 export const parseModelOption = (model?: string) => {
-  const { provider, name } = model?.match(/(?<provider>[^:]+)(:(?<name>(\S+)))?/)?.groups ?? {};
+  const { provider, name } =
+    (model || process.env.MODEL)?.match(/(?<provider>[^:]+)(:(?<name>(\S+)))?/)?.groups ?? {};
 
   return { provider, name };
 };
@@ -207,7 +211,7 @@ export async function runWithAIGNE(
       }
 
       const model = await loadModel(
-        availableModels,
+        availableModels(),
         {
           ...parseModelOption(options.model),
           temperature: options.temperature,
@@ -240,7 +244,7 @@ export async function runWithAIGNE(
 
         await runAgentWithAIGNE(aigne, agent, {
           ...options,
-          outputKey,
+          outputKey: outputKey || options.outputKey,
           chatLoopOptions,
           modelOptions,
           input,
@@ -304,6 +308,7 @@ export async function runAgentWithAIGNE(
 
     await runChatLoopInTerminal(userAgent, {
       ...chatLoopOptions,
+      outputKey,
     });
 
     return;