+++ packages/core/src/aigne/context.ts
@@ -1,6 +1,7 @@
-import type { AIGNEObserver } from "@aigne/observability";
-import type { Span } from "@opentelemetry/api";
+import type { AIGNEObserver } from "@aigne/observability-api";
 import { SpanStatusCode, context, trace } from "@opentelemetry/api";
+import type { Span } from "@opentelemetry/api";
+import type { ReadableSpan } from "@opentelemetry/sdk-trace-base";
 import equal from "fast-deep-equal";
 import { Emitter } from "strict-event-emitter";
 import { v7 } from "uuid";
@@ -41,7 +42,7 @@ import {
   isNil,
   omit,
 } from "../utils/type-utils.js";
-import type { Args, Listener, TypedEventEmitter } from "../utils/typed-event-emtter.js";
+import type { Args, Listener, TypedEventEmitter } from "../utils/typed-event-emitter.js";
 import {
   type MessagePayload,
   MessageQueue,
@@ -407,9 +408,11 @@ export class AIGNEContext implements Context {
       options.memories = undefined;
     }
 
+    const newContext = options?.newContext === false ? this : this.newContext();
+
     return this.internal.messageQueue.publish(topic, {
       ...toMessagePayload(payload),
-      context: this,
+      context: newContext,
     });
   }) as Context["publish"];
 
@@ -463,6 +466,26 @@ export class AIGNEContext implements Context {
           span.setAttribute("input", JSON.stringify(input));
           span.setAttribute("agentTag", agent.tag ?? "UnknownAgent");
 
+          try {
+            span.setAttribute("userContext", JSON.stringify(this.userContext));
+          } catch (_e) {
+            logger.error("parse userContext error", _e.message);
+            span.setAttribute("userContext", JSON.stringify({}));
+          }
+
+          try {
+            span.setAttribute("memories", JSON.stringify(this.memories));
+          } catch (_e) {
+            logger.error("parse memories error", _e.message);
+            span.setAttribute("memories", JSON.stringify([]));
+          }
+
+          await this.observer?.traceExporter
+            ?.upsertInitialSpan?.(span as unknown as ReadableSpan)
+            .catch((err) => {
+              logger.error("upsertInitialSpan error", err?.message || err);
+            });
+
           break;
         }
         case "agentSucceed": {
@@ -475,17 +498,29 @@ export class AIGNEContext implements Context {
             span.setAttribute("output", JSON.stringify({}));
           }
 
-          span.setStatus({ code: SpanStatusCode.OK, message: "Agent succeed" });
+          span.setStatus({ code: SpanStatusCode.OK });
+
+          await this.observer?.traceExporter
+            ?.upsertInitialSpan?.(span as unknown as ReadableSpan)
+            .catch((err) => {
+              logger.error("upsertInitialSpan error", err?.message || err);
+            });
+
           span.end();
-          await this.observer?.traceExporter?.forceFlush?.();
 
           break;
         }
         case "agentFailed": {
           const { error } = args[0] as ContextEventMap["agentFailed"][0];
           span.setStatus({ code: SpanStatusCode.ERROR, message: error.message });
+
+          await this.observer?.traceExporter
+            ?.upsertInitialSpan?.(span as unknown as ReadableSpan)
+            .catch((err) => {
+              logger.error("upsertInitialSpan error", err?.message || err);
+            });
+
           span.end();
-          await this.observer?.traceExporter?.forceFlush?.();
 
           break;
         }