+++ packages/observability/api/third.d.ts
@@ -1,8 +1,0 @@
-declare module "vite-plugin-blocklet";
-
-declare namespace Express {
-  interface Request {
-    user?: import("@blocklet/sdk/lib/util/login").SessionUser;
-    isBlocklet?: boolean;
-  }
-}