+++ packages/observability/src/env.d.ts
@@ -1,4 +1,0 @@
-declare const blocklet: import("@blocklet/sdk").WindowBlocklet | undefined;
-declare module "@arcblock/ux/*";
-declare module "@arcblock/did-connect/*";
-declare module "@arcblock/ux/*";